+++ docs/source/changelog.md
@@ -40,6 +40,11 @@ __nodepy-pm (PPYM post v0.0.17)__
   modules are now installed nearby Python's `site-packages` and Scripts
   are placed into Python's script directory
 - Development has been merged into the Node.py repository itself
+- Remove `Context.register_extension()` and `Context.get_extension()`
+- Add instead `Context.loaders` and `Context.get_loader()`
+- Add `BaseLoader` class with `suggest_try_files()`, `can_load()` and `load()`
+  methods
+- Add `PythonLoader(support_require_unpack_syntax)` argument
 
 ### v0.0.19